linux-yocto-dev: Fix missing stat64 system call
The planned upstream release of 32-bit RISC-V glibc will support statx,
so this patch will not be required in the kernel.

Unfortunately the current 32-bit RISC-V glibc fork does not implement
statx. To allow newer kernels to boot into userspace let's apply this
patch that can be dropped when the glibc port is merged.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>