Add meta-python before meta-multimedia
meta-multimedia depends on meta-python.  I'm not sure why it's
necessary to add all the dependent layers in order, but if I don't do
this I get a bunch of exceptions during setup.sh that result in no
meta-python in my bblayers.conf and thus a non-functional bitbake.

setup.sh was already explicitly adding meta-python, so this patch just
adds it before meta-multimedia.  That fixes the errors on my side.

Signed-off-by: Palmer Dabbelt <palmer@sifive.com>