linux-beaglev-dev: provide Kernel for Beagle-V Ahead
The Kernel follows latest development of the Linux upstream Kernel. For
now, use a separate recipe to isolate Beagle-V specific versions and
inclusion of pending upstream packages.

The defconf was taken from: https://openbeagle.org/beaglev-ahead/xuantie-ubuntu

The recipe also provides a custom function to package the boot.ext4
deployment artifact, which is needed for fastboot based deployment of
the device's eMMC.