star64: fix u-boot compilation
Make sure that the do_install step is applied to the correct boards.
do_install:append:jh7110 was executed for star64 also (since it is
based on the jh7110 SoC as well), however the installed file is only
used for visionfive2 board, failing the build.

This change ensures that this file is only installed for visionfive2 board,
and not for star64.