visionfive2-pvr-graphics: fix installation with SysV init
Fix the installation of this recipe when using SysV init system by
fixig the "if" condition that checks the presence of systemd in DISTRO_FEATURES -
beforehand it was always evaluated as "true".

Beside this the required folders are also created when they should be:
in case of SysV init the /usr/bin folder stays empty (and bitbake complains),
so it is only created when systemd is used.

Similarly, /etc/init.d is used only by SysV init system, but the recipe already
handles it when systemd is used. However if the folder is not installed explicitly,
the do_rootfs task fails.