recipes-bsp/firmware/firmware-rcpu-orangepi-rv2.bb: Add RCPU firmware recipe for OrangePi RV2
- esos.elf is the firmware for the RCPU (Real-Time CPU), responsible for initializing
some hardware modules and forwarding HDMI Audio interrupts. It is dependent on the Linux kernel
and the system will not boot without it and needs to be installed in the initramfs /lib/firmware directory.

- Marked to skip certain QA checks.

- Only compatible with OrangePi RV2.

Note: esos.elf is a closed source firmware binary and doesn't have an explicit LICENSE/COPYING file.
It isn't available in any official Orange Pi repository, but was obtained from:

https://github.com/orangepi-xunlong/orangepi-build/blob/next/external/packages/bsp/ky/usr/lib/firmware/esos.elf

Due to its closed and undocumented nature, the recipe sets LICENSE = "CLOSED".

Signed-off-by: Alper Ak <alperyasinak1@gmail.com>