linux-orangepi: Use alldefconfig without leading -- in KCONFIG_MODE
In kernel-yocto.bbclass, the shell case statement uses the *alldefconfig pattern. Here, the asterisk (*) acts as a wildcard and matches any characters before alldefconfig.

Therefore, any value ending with alldefconfig including non standard forms like --alldefconfig will be matched and processed.

But, it makes more sense to stick to the documentation and update it in this way.

https://docs.yoctoproject.org/ref-manual/variables.html#term-KCONFIG_MODE

Signed-off-by: Alper Ak <alperyasinak1@gmail.com>