add support for OrangePi RV2 with mainline Linux
This introduces a new "orangepi-rv2-mainline" machine
and a new "linux-orangepi-mainline" kernel recipe.

This machine definition and kernel recipe are not even trying to share code
with the original ones for "orangepi-rv2", because they are ultimately meant
to replace them.

The U-Boot version remains the vendor one for the moment.

Signed-off-by: Michael Opdenacker <michael.opdenacker@rootcommit.com>