nezha-common: set KERNEL_IMAGETYPE to 'Image'
KERNEL_IMAGETYPE is derived from ARCH_DEFAULT_KERNELIMAGETYPE by
default, which is 'zImage' in oe-core for every architecture other than
x86. However, building for the nezha with this results in a "No rule to
make target 'zImage'" error for the kernel. Several of the other boards
in the meta-riscv layer set KERNEL_IMAGETYPE = "Image", so do the same
here.

Signed-off-by: Trevor Gamblin <tgamblin@baylibre.com>