README.md: add Contributing section, fix maintainer email format
The yocto-check-layer script (which must pass as part of the
requirements for Yocto Project Compatible Status) checks the README file
for occurrences of the word "patch", presumably as a basic way of
checking that it contains some instructions on how to submit patches.

For comparison, oe-core's README.md has the following entry under the
"Contributing" header that passes the check:

|As a quick guide, patches should be sent to
|openembedded-core@lists.openembedded.org
|The git command to do that would be:

Since meta-riscv is currently hosted on GitHub, add a "Contributing"
section that tells the reader to submit changes via PR and to use the
issue tracker for comments (like meta-altera).

Also fix the maintainer's email address format so that it passes the
yocto-check-layer email regex test.

Signed-off-by: Trevor Gamblin <tgamblin@baylibre.com>