bananapi-f3: new MACHINE
For the Banana PI F3 board with the SpacemiT K1 SoC:
https://docs.banana-pi.org/en/BPI-F3/BananaPi_BPI-F3

This board is supported here with the mainline U-Boot
and Linux kernel. Only U-BOOT SPL (with DRAM training)
still originates from the SoC vendor tree.

Mainline U-Boot and kernel support is still limited,
as there is no MMC support yet.

However, this new MACHINE exists to make it easier to use
and expand mainline U-Boot and Linux on this board and SoC.

U-Boot's boot.cmd file is currently a placeholder, and will
be used when U-Boot is able to load an environment from MMC.

See docs/bananapi-f3.md for details about building and
booting an image.

Signed-off-by: Michael Opdenacker <michael.opdenacker@rootcommit.com>